Polish previous changes and refined serial open process.

+++ conveyer_serial_slave/conveyer_serial_slave.ino
@@ -68,7 +68,7 @@ if (Serial.available() >= 3){
     int speed = steps * 300/interval;
 
     // Relay confirmation data was received
-    Serial.write(STEPCODE);
+    Serial.write((byte)STEPCODE);
     Serial.write(steps);
   }
